fix(metrics): promote iwssim cov_finalize to f64 + tighten parity tolerances
Root-cause fixes that let the parity gates tighten without relaxing
correctness:

1. **iwssim cov_finalize: f32 sum -> f64 sum** (kernels/cov.rs +
   pipeline.rs). The per-cell finalize kernel summed 16384 f32 partials
   in f32, hitting a sqrt(N)*epsilon ~7.7e-6 per-cell floor that
   propagated through eigendecomp and the |wmcs|^beta product to
   ~2-3e-4 final score drift in multi-strip parity. Cubecl supports
   f64 directly; the partials buffer stays f32 (per-thread small-N
   sums) but the cross-thread reduction now widens to f64. cu buffer
   becomes f64 (100 cells * 8 B = 800 B per scale - negligible).
   read_cu_raw + scale_cu signatures updated accordingly. Multi-strip
   max rel drift drops from 5e-4 to 3.6e-6 (~140x tighter floor).

2. **iwssim strip_parity** (tests/strip_parity.rs): STRIP_VS_WHOLE_REL
   5e-4 -> 1e-5 (3x margin over measured 3.6e-6). STRIP_SINGLE_REL
   1e-4 -> 1e-7 (single-strip degenerate is now bit-exact 0.0).
   STRIP_VS_STRIP_REL 1.5e-3 -> 5e-5. Module docs rewritten to
   document the f64-finalize design.

3. **ssim2 modes_agree_on_jpeg_corpus** (tests/ssim2_skipmap_audit.rs):
   conditional tolerance on the fast-reduction feature flag.
   - With fast-reduction (default): 5e-5 gate (Atomic<f32>::fetch_add
     reorder noise, measured max 1.03e-5).
   - Without fast-reduction (portable, deterministic finalize): 1e-6
     gate (measured 0.0 exactly).
   Both paths now get coverage at their actual precision floor instead
   of the loosest path's gate.

4. **zensim cpu_parity / extended_parity**: tightened the loose abs
   bounds that were 70x above the measured floor. New gates:
   identical case 5e-2 -> 2e-3; noisy_gradient_64 abs gates tightened
   from 5e-3 to 2e-3 / 1e-3 / 2e-4 (basic / peak / l8); checkerboard
   128 abs gates tightened from 5e-3/3e-3/2e-3 to 2e-3/5e-4/5e-4;
   with_iw[0..300] vs Extended divergence gate 5e-3 -> 1e-9
   (measured bit-identical). Documented why rel cannot tighten
   without f64-ifying the mu/sigma/cov maps in the fused per-cell
   kernel (an expensive memory-bandwidth tradeoff).

Per-target outcome: ssim2 modes_agree 5e-5 -> 1e-6 (without
fast-reduction) via deterministic portable reduction; iwssim
multi-strip 5e-4 -> 1e-5 via f64 cov finalize; zensim cpu_parity
abs gates tightened to measured floor (rel band documented as
irreducible without map-level f64 promotion).

All existing parity tests pass with the new tighter gates on cuda.
wgpu + cpu backends build cleanly with the same code changes.

+++ crates/iwssim-gpu/src/kernels/cov.rs
@@ -666,15 +666,27 @@ pub fn cov_accum_with_parent_kernel(
 /// `CubeDim::new_1d(1)` (one thread per cube — minimal but matches the
 /// shape of reduction.rs's `finalize_kernel`; could parallelize per-cube
 /// later if profiling shows it matters).
+///
+/// Accumulates in **f64** (the `partials` buffer is `f32` but each
+/// addend is widened to f64 before the running sum). Per-cell error of
+/// the prior f32 finalizer was bounded by √N · ε_f32 ≈ √16384 · 6e-8 ≈
+/// 7.7e-6 (relative to the cell magnitude), which propagated through
+/// the per-strip f64 host accumulator, eigendecomp, and Π|wmcs|^β to
+/// ~2-3e-4 final JOD drift in multi-strip parity tests. Promoting the
+/// cross-thread sum to f64 reduces the per-cell floor to ~ε_f64 ≈
+/// 1e-15 and lets the multi-strip parity gate tighten to the 1e-5 band
+/// (driven by the f32 per-thread accumulator, which still sums a small
+/// number of products per thread). Per-cell f64 cells cost 800 bytes
+/// per scale (5×800 = 4 KB total) — negligible.
 #[cube(launch_unchecked)]
-pub fn cov_finalize_kernel(partials: &Array<f32>, cu: &mut Array<f32>, n_threads: u32) {
+pub fn cov_finalize_kernel(partials: &Array<f32>, cu: &mut Array<f64>, n_threads: u32) {
     let cell = CUBE_POS_X;
     let n_t = n_threads as usize;
     let base = (cell as usize) * n_t;
-    let mut s = 0.0_f32;
+    let mut s = 0.0_f64;
     let mut k: u32 = 0;
     while k < n_threads {
-        s += partials[base + (k as usize)];
+        s += partials[base + (k as usize)] as f64;
         k += 1;
     }
     cu[cell as usize] = s;
